Refactor logic for updating locations
