+++ src/main/java/org/opensrp/repository/postgres/LocationRepositoryImpl.java
@@ -482,16 +482,6 @@ public class LocationRepositoryImpl extends BaseRepositoryImpl<PhysicalLocation>
 	}
 
 	@Override
-	public boolean isGeometryCoordsEqual(PhysicalLocation newEntity, PhysicalLocation existingEntity) {
-		if (newEntity == null || existingEntity == null) {
-			return false;
-		}
-		JsonElement newGeometryCoordsElement = JsonParser.parseString(newEntity.getGeometry().getCoordinates().toString());
-		JsonElement existingGeometryCoordsElement = JsonParser.parseString(existingEntity.getGeometry().getCoordinates().toString());
-		return newGeometryCoordsElement.equals(existingGeometryCoordsElement);
-	}
-
-	@Override
 	public PhysicalLocation findLocationByIdentifierAndStatus(String identifier, String status, boolean returnGeometry) {
 		LocationMetadataExample locationMetadataExample = new LocationMetadataExample();
 		locationMetadataExample.createCriteria().andGeojsonIdEqualTo(identifier)
